+++ app/Filament/Resources/Categories/Pages/ListCategories.php
@@ -2,8 +2,12 @@
 
 namespace App\Filament\Resources\Categories\Pages;
 
+use App\Actions\CategoryActions\CreateCategoryAction;
 use App\Filament\Resources\Categories\CategoryResource;
+use App\Filament\Resources\Categories\Schemas\CategoryForm;
+use Daljo25\FilamentTablerIcons\Enums\TablerIcon;
 use Filament\Actions\CreateAction;
+use Filament\Notifications\Notification;
 use Filament\Resources\Pages\ListRecords;
 
 class ListCategories extends ListRecords
@@ -13,7 +17,21 @@ class ListCategories extends ListRecords
     protected function getHeaderActions(): array
     {
         return [
-            CreateAction::make(),
+            CreateAction::make()
+                ->form(CategoryForm::getComponents())
+                ->label(__('Create Category'))
+                ->icon(TablerIcon::Plus)
+                ->action(function (array $data) {
+                    CreateCategoryAction::execute($data);
+                }),
         ];
+    }
+
+    public function notifications(): void
+    {
+        Notification::make()
+            ->title(__('Category created successfully'))
+            ->success()
+            ->send();
     }
 }