+++ app/Filament/Resources/Categories/Tables/CategoriesTable.php
@@ -2,10 +2,14 @@
 
 namespace App\Filament\Resources\Categories\Tables;
 
+use Daljo25\FilamentTablerIcons\Enums\TablerIcon;
 use Filament\Actions\BulkActionGroup;
 use Filament\Actions\DeleteBulkAction;
 use Filament\Actions\EditAction;
 use Filament\Actions\ViewAction;
+use Filament\Tables\Columns\ColorColumn;
+use Filament\Tables\Columns\IconColumn;
+use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
 
 class CategoriesTable
@@ -14,7 +18,27 @@ class CategoriesTable
     {
         return $table
             ->columns([
-                //
+                TextColumn::make('name')
+                    ->label(__('Name'))
+                    ->searchable()
+                    ->sortable(),
+                TextColumn::make('slug')
+                    ->label(__('Slug'))
+                    ->searchable()
+                    ->sortable(),
+                ColorColumn::make('color')
+                    ->label(__('Color'))
+                    ->searchable()
+                    ->sortable(),
+                IconColumn::make('icon')
+                    ->icon(fn ($state) => $state ?? TablerIcon::AccessPoint)
+                    ->label(__('Icon'))
+                    ->searchable()
+                    ->sortable(),
+                TextColumn::make('sort_order')
+                    ->label(__('Sort Order'))
+                    ->searchable()
+                    ->sortable(),
             ])
             ->filters([
                 //